Make ring slightly bigger to cover circle
Because the ring size is exactly the same size as the circle, the circle
can show around the outside of the ring (screen resolution, maybe?).  I
made the ring slightly bigger.

I also adjusted the margin of the circle so that the circle remains
centered in the ring (the top edge is aligned, but the height of the
circle and the ring are different because the ring is slightly bigger).