Allow animation box to grow
The problem was that the yellow accent box in the top left was shorter than its specified width.  Allowing the animation box to grow solved this problem.

The flex-grow value is a "flex grow factor".  It specifies how much of
the "slack" space remaining gets assigned to the element:  The higher
the factor, the more space gets allocated to that element.