Add custom run button
Created an animated run button.  This also fixed a styling problem.

Before the button, the animation started as soon as the page was loaded.
However, I noticed that the right end point of the animation was
slightly off-center to the ring.  After probing around, it turns out
that the animation starts before the custom font can be loaded in.  This
caused the ring to be shifted to the left to accomodate the new font.
However, the right bound wasn't updated, causing the circle to land
off-center to the right ring.

I tried using useState and useEffect to update the rightBound, but it
didn't change anything: Once the animation starts it seems to ignore the
updated right bound.

The only solution using useState and useEffect was to allow useEffect to
watch the x keyframe array to the animation.  But the reason it worked
was that useEffect was constantly updating because it was watching an
array.  You're supposed to watch a value in the array, not the array
itself.  As far as I understand, React can't tell if a value in the
array is updated or not.

In the end, I decided to add a run button.  When you click the button,
the right bound gets updated and _then_ the animation is allowed to
start.