Implement simple collision detection for trip counter
Initially, I was thinking about a traditional 2d collision detection
system like here:

<https://developer.mozilla.org/en-US/docs/Games/Techniques/2D_collision_detection>

But after thinking about it, my app only needs to know if the flying
circle has passed certain coordinates.  The flying circle is constrained
to a simple flight path and only hits the destination rings from one
side.  If the flying circle had to approach from any side and was
allowed an arbitrary flight path, I would have to switch to traditional
collision detection.

The current system for incrementing the trip counter is based on
detecting when the flying circle transitions between three different
rectangular regions: the region including the left ring at the right
x-coordinate of the ring, the region between the two rings, and the
region including the right ring at the left x-coordinate of the right
ring.

Mathematically, the simple system detects when the flying circle moves
past a vertical line located at the right edge of the left ring and a
vertical line located at the left edge of the right ring.

But it only registers a collision when the flying circle is moving _out_
of the region between the two rings.  If I didn't do this, the trip
counter would register a trip right when the animation started because
the flying circle starts outside this region.